优化

+++ src/Services/SurveillanceDeviceService.php
@@ -63,17 +63,19 @@ class SurveillanceDeviceService extends AdminService
     {
         parent::saved($model, $isEdit);
         $request = request()->all();
-        $data = [
-            'enterprise_id' => $request['enterprise_id'],
-            'facility_id' => $request['facility_id'],
-            'device_id' => $model->id,
-        ];
-        admin_transaction(function () use ($data) {
-            if ($data['device_id']) {
-                EnterpriseFacilityDevice::query()->where($data)->delete();
-            }
-            EnterpriseFacilityDevice::query()->insert($data);
-        });
+        if ($model->id && !empty($request['enterprise_id']) && !empty($request['facility_id'])) {
+            $data = [
+                'enterprise_id' => $request['enterprise_id'],
+                'facility_id' => $request['facility_id'],
+                'device_id' => $model->id,
+            ];
+            admin_transaction(function () use ($data) {
+                if ($data['device_id']) {
+                    EnterpriseFacilityDevice::query()->where($data)->delete();
+                }
+                EnterpriseFacilityDevice::query()->insert($data);
+            });
+        }
     }
 
     /**
